move dropdownOptions from common to table
