+++ frontend/src/components/Lists/Table/Dialogs/Employees/EmployeeDialogContent.jsx
@@ -3,7 +3,7 @@ import TextField from '@material-ui/core/TextField';
 import Autocomplete from '@material-ui/lab/Autocomplete';
 import PropTypes from 'prop-types';
 import React from 'react';
-import { employeeTypes } from '../../../../../common/dropdownOptions';
+import { employeeTypes } from '../../dropdownOptionsEmployee';
 
 function EmployeeDialogContent(props) {
   const { employee, setEmployee, newEmployee } = props;
